Add two directives called CrowdCertPath and CrowdSSLVerifyPeer.
I wrote a patch for my environment to extend the module to have some
additional configuration directives to give you the option of specifying a
path to curl's SSL certificate store or to allow you to choose to disable
peer validation altogether. The module defaults to verifying peer SSL
certificates but at least gives the end user the option of consciously
choosing to disable the verification if they want to or need to (i.e. in the
case of wildcard certs where even with the cert imported in curl the peer
fails hostname validation).